+++ src/Delizious.Toolkit.Test/MatchSpec.cs
@@ -222,7 +222,6 @@ namespace Delizious
             }
         }
 
-
         public sealed class NotEqualEqualityComparer
         {
             [Fact]
@@ -327,6 +326,37 @@ namespace Delizious
                 yield return DataTheory(false, "Test", "TEST",  StringComparer.Ordinal);
                 yield return DataTheory(true,  "Test", "TEST",  StringComparer.OrdinalIgnoreCase);
                 yield return DataTheory(true,  "Test", "test",  StringComparer.OrdinalIgnoreCase);
+            }
+        }
+
+        public sealed class NotEqualToComparable
+        {
+            [Fact]
+            public void Throws_exception_when_reference_value_is_null()
+            {
+                Assert.Throws<ArgumentNullException>(() => Match.NotEqualTo<string>(null));
+            }
+
+            [Theory]
+            [MemberData(nameof(MatchesTheories))]
+            public void Matches(bool expected, int reference, int value)
+            {
+                var subject = Match.NotEqualTo(reference);
+
+                var actual = subject.Matches(value);
+
+                Assert.Equal(expected, actual);
+            }
+
+            public static IEnumerable<object[]> MatchesTheories()
+            {
+                yield return DataTheory(false, 1,  1);
+                yield return DataTheory(true,  1,  0);
+                yield return DataTheory(true,  0,  1);
+                yield return DataTheory(false, 0,  0);
+                yield return DataTheory(true,  -1, 0);
+                yield return DataTheory(true,  0,  -1);
+                yield return DataTheory(false, -1, -1);
             }
         }